Feat: 딜 스코어 double 변경, 소숫점 버리고 Int형으로 response

+++ src/main/java/com/cherrypick/backend/domain/deal/entity/Deal.java
@@ -73,7 +73,7 @@ public class Deal {
     private List<DealTag> dealTags = new ArrayList<>();
 
     @Column(nullable = false)
-    private int dealScore = 0;
+    private double dealScore = 0.0;
 
     @Builder.Default
     private Boolean isDelete = false;
